TrashManager: Refactor from vector to map structure
Since using vector as a dynamic array is completely fine, and also suggested,
however, vector does not check duplication of entries. Which we need to make another method to search all-around for duplication.

That kind of operation is pretty much exhaustive - method to approach goals.
Therefore refactor vector to map. It does automatically do it for us.