TrashManager: Add support for removing duplicated entry on TrashData
Since we do not need to think of duplicated entry, because "duplicated"
means there is an entry on our trashdata, but it is NOT actually on
Trash.

The main reason why trashdata exists is: To check/restore accidentally
files via this program[command]. If the entry is NOT exists in Trashcan,
it is kinda abundant.

WARNING:
This commit includes entry modification. This executing commit[or feature] will
remove all of duplicated entry[where duplication bool is true] and
cannot be undone.